Fix AP issues reoccurring in SDK5 games on DSi
Fixes #1109

+++ retail/bootloaderi/source/arm7/ips.c
@@ -9,7 +9,7 @@
 #include "locations.h"
 #include "tonccpy.h"
 
-extern u32 consoleModel;
+extern u8 consoleModel;
 extern bool dsiModeConfirmed;
 extern bool extendedMemoryConfirmed;
